+++ src/main/java/org/example/gametgweb/characterSelection/infrastructure/persistence/repository/JpaUnitRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
  * Позволяет выполнять CRUD-операции и искать юниты по имени.
  */
 @Repository
-public interface JpaUnitRepository extends JpaRepository<UnitEntity, Integer> {
+public interface JpaUnitRepository extends JpaRepository<UnitEntity, Long> {
 
     /**
      * Находит юнита по его имени.
